pyproject.toml
[tool.black]
line-length = 110
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = "apps,quizhero-api,tests,utils"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "quizhero_api.settings"
addopts = "-vv --reuse-db --tb=native --cov=apps"
pytestenvvars__env_files = [
    ".env",
]
pytestenvvars__dont_randomize_envvars = [
    "DATABASE_URL",
]

[tool.poetry]
name = "quizhero-api"
version = "0.1.0"
description = ""
authors = ["MatheusBLopes <matheusblopes10@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
prettyconf = "^2.2.1"
gunicorn = "^20.1.0"
gevent = "^21.12.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pre-commit = "^2.20.0"
safety = "^2.0.0"
pytest-cov = "^3.0.0"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^4.5.2"
pytest-envvars = "^1.2.1"
pytest-randomly = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
